ugtrain: read multiple pages from the dynmem FIFO
In the second single-player challenge in Warzone 2100 there are
806 Structure objects which are created at once. This is too much
for a single page read and leads to instability and victim memory
r/w errors. So read until the FIFO buffer is empty.

+++ src/fifoparser.h
@@ -32,8 +32,8 @@ struct post_parse {
 	void	*argp;
 };
 
-i32 read_dynmem_buf (list<CfgEntry> *cfg, void *argp, i32 ifd, int pmask, bool reverse,
-		     void (*mf)(list<CfgEntry> *, struct post_parse *, void *,
-				void *, size_t, void *, void *),
-		     void (*ff)(list<CfgEntry> *, void *, void *));
+ssize_t read_dynmem_buf (list<CfgEntry> *cfg, void *argp, i32 ifd, int pmask, bool reverse,
+			 void (*mf)(list<CfgEntry> *, struct post_parse *, void *,
+				    void *, size_t, void *, void *),
+			 void (*ff)(list<CfgEntry> *, void *, void *));
 #endif
